Update 5.9 PATTERN: DESIGN FOR IDEMPOTENCY.md
I propose to delete a very dangerous sentence.

You recommend using Redis for API idempotency control. I will assume the context is that there is some other database for the final operation, and that Redis is then intended as a cache in front?

First off, any naive implementation including Redis will leave the code wide open for race conditions. One must care to use Redis transactions and the single-threaded nature of Redis in order to not be vulnerable to races. This in itself is a weakness in such a design.

But then comes the main problem: What if the first request gets stored in Redis, but then power disappears and processing stops before one hits the main database? Or the main database is unavailable, or whatever. This is exactly where the retry from the client, the at-least-once-principle, is helpful! But, it will then get stopped by the request already being present in Redis (but, the operation hasn't happened in the main database -- and so one has a bug!).

Idempotency need to happen *all the way in to the main database*. (For instance, in SQL, just store the idempotency ID and use a unique constraint.) Otherwise, there is no point in doing it!

Second problem: Redis is typically a *cache*, so that data can disappear and is not guaranteed to be durable for the given duration.

Idempotency that only works 99.9% of the time is entirely worthless, especially since one incident often triggers another (some failure triggers a lot of retries, which overloads Redis, someone restarts Redis and wipes it out ...). One needs a solution that works 100% of the time.

Using a Redis cache for idempotency control is simply wrong. In an interview setting, I would not hire someone who made such a suggestion; at least unless they were junior and very willing to learn from more experienced people...

I am not sure how to find a polite way of saying this. But: With the kind of advice you give here -- then perhaps you should not share advice to others about how to architect their applications. You need to learn the basics yourself, before you can teach others.